requirements.txt
pandas
requests
pytest
pytest-html
pyperclip
lxml
xlrd
beautifulsoup4
html5lib
pyyaml 